requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data Sources
yfinance>=0.2.28
alpha-vantage>=2.3.1
tiingo>=0.14.0
# eodhistoricaldata>=1.0.0  # Optional - package not available on PyPI
quandl>=3.7.0

# Technical Analysis
ta-lib>=0.4.28
pandas-ta>=0.3.14b0
ta>=0.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0
mplfinance>=0.12.10

# Progress Bars
tqdm>=4.66.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Statistics
statsmodels>=0.14.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Reporting
reportlab>=4.0.0
jinja2>=3.1.0

# Streaming (NEW - for real-time analysis)
websockets>=12.0
aiohttp>=3.9.0
alpaca-trade-api>=3.0.0  # Optional: for Alpaca streaming